+++ PGIsertAdministratorRevistas5642.php
@@ -1,3 +1,15 @@
+<?php
+  session_start();
+
+
+  if(empty($_SESSION['adm']))
+  {
+    echo ('<meta http-equiv="refresh"content=0;"index.php">');
+  }
+  else
+  {
+    
+?>
 <!DOCTYPE html>
 <html lang="pt-br">
 <head>
@@ -208,17 +220,6 @@
             </li>
       </ul>
 
-      
-
-        <!-- Pesquisa do site-->
-        <form class="form-inline mt-2 mt-md-0">
-            <input class="form-control mr-sm-2" type="text" placeholder="Pesquisar" aria-label="Search">
-            <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Pesquisar</button>
-        </form>
-
-
-
-
     </div>
 </div>
 
@@ -287,13 +288,13 @@
 								$extensao = pathinfo($arquivo['name'], PATHINFO_EXTENSION);
 								$novo_nome = md5(uniqid($arquivo['name'])).".".$extensao;
 
-								move_uploaded_file($_FILES['imgboxRevista']['tmp_name'], $_UP['pasta'].$novo_nome);
+								
 
 								$sql_code="INSERT INTO `tb_revistas` (`COD_REVISTA`, `NOME_REVISTA`, `EDITORA_REVISTA`, `IMG_REVISTA`) VALUES ('$codrevista', '$nomerevista', '$edirarevista', '$novo_nome');";
 
 								if(mysqli_query($conn, $sql_code))
 								{
-									
+									move_uploaded_file($_FILES['imgboxRevista']['tmp_name'], $_UP['pasta'].$novo_nome);
 									echo "<script>alert('Arquivo enviado com susesso');</script>";
 								}
 								else
@@ -334,27 +335,47 @@
 					{
 						if($delatarrevista != null)
 						{
-              $sql_consultaimg = "SELECT IMG_REVISTA FROM tb_revistas WHERE COD_REVISTA LIKE '$delatarrevista'";
+              $sql_consultacod = "SELECT COD_REVISTA FROM tb_revistas WHERE COD_REVISTA LIKE '$delatarrevista'";
 
-              $resultado_consultaimg = mysqli_query($conn, $sql_consultaimg);
+              $resultado_consultacod = mysqli_query($conn, $sql_consultacod);
 
-              $row_consultaimg = mysqli_fetch_assoc($resultado_consultaimg);
+              $row_consultacod = mysqli_fetch_assoc($resultado_consultacod);
 
-              $img_cod = $row_consultaimg['IMG_REVISTA'];
+              $cod_consultado = $row_consultacod['COD_REVISTA'];
 
-							$sql_code ="DELETE FROM `tb_revistas` WHERE `tb_revistas`.`COD_REVISTA` = '$delatarrevista'";
+
+              if($cod_consultado != null)
+              {
+                $sql_consultaimg = "SELECT IMG_REVISTA FROM tb_revistas WHERE COD_REVISTA LIKE '$delatarrevista'";
+
+                $resultado_consultaimg = mysqli_query($conn, $sql_consultaimg);
+
+                $row_consultaimg = mysqli_fetch_assoc($resultado_consultaimg);
+
+                $img_cod = $row_consultaimg['IMG_REVISTA'];
+
+  							$sql_code ="DELETE FROM `tb_revistas` WHERE `tb_revistas`.`COD_REVISTA` = '$delatarrevista'";
 
 								if(mysqli_query($conn, $sql_code))
 								{
-									unlink ("upload/" . $img_cod );
+									if(file_exists("upload/" . $img_cod))
+                  {
+                    unlink ("upload/" . $img_cod ) ;
+                  }
+                  else
+                  {
+                    echo"";
+                  }
+                  
 									echo "<script>alert('Arquivo Deletado com susesso');</script>";
-								}
-								else
-								{
-									echo '<h3 style="color: red;">';
-									echo 'Numero não existe';
-									echo '</h3>';
-								}
+								}		
+              }
+              else
+              {
+                echo '<h3 style="color: red;">';
+                echo 'Código não existe';
+                echo '</h3>';
+              }
 						}
 						else
 						{
@@ -363,7 +384,7 @@
 							echo '</h3>';
 						}
 					}
-
+        }
 				?>
 			</div>
 		</form>
